+++ app/components/Reporting/Tabs.tsx
@@ -10,11 +10,18 @@ const StyledNav = styled.nav`
 const Tabs = () => {
   const router = useRouter();
   const gcpeHref = '/analyst/reporting/gcpe';
+  const manageReportsHref = '/analyst/reporting/manage-reports';
   // this a bare page to handle any future reporting tabs
   return (
     <StyledNav>
       <StyledTab href={gcpeHref} selected={router?.pathname.includes(gcpeHref)}>
         GCPE
+      </StyledTab>
+      <StyledTab
+        href={manageReportsHref}
+        selected={router?.pathname.includes(manageReportsHref)}
+      >
+        Manage Reports
       </StyledTab>
     </StyledNav>
   );