+++ app/lib/helpers/useEmailNotification.ts
@@ -1,6 +1,6 @@
 import * as Sentry from '@sentry/nextjs';
 
-const useHHCountUpdateEmail = () => {
+const useEmailNotification = () => {
   const notifyHHCountUpdate = async (
     newData: any,
     oldData: any,
@@ -48,7 +48,30 @@ const useHHCountUpdateEmail = () => {
     });
   };
 
-  return { notifyHHCountUpdate };
+  const notifyDocumentUpload = async (applicationId: string, params: any) => {
+    fetch('/api/email/notifyDocumentUpload', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({
+        applicationId,
+        host: window.location.origin,
+        params: {
+          ...params,
+          timestamp: new Date().toLocaleString(),
+        },
+      }),
+    }).then((response) => {
+      if (!response.ok) {
+        Sentry.captureException({
+          name: `Error sending email to notify ${params.documentType} upload`,
+          message: response,
+        });
+      }
+      return response.json();
+    });
+  };
+
+  return { notifyHHCountUpdate, notifyDocumentUpload };
 };
 
-export default useHHCountUpdateEmail;
+export default useEmailNotification;